requirements.txt
sounddevice
numpy
scipy
mistralai
python-dotenv
pynput
pystray
Pillow
PyGObject
